Work around hard crash with Iris
This causes Iris to no longer show any rendering, but it's better than a hard crash.
