+++ src/main/java/dev/kir/sync/client/render/CustomRenderLayer.java
@@ -1,7 +1,10 @@
 package dev.kir.sync.client.render;
 
+import dev.kir.sync.compat.iris.IrisCustomRenderLayer;
+import dev.kir.sync.compat.iris.IrisRenderLayer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.RenderPhase;
 import net.minecraft.client.render.VertexFormat;
@@ -28,12 +31,20 @@ public final class CustomRenderLayer extends RenderLayer {
     }
 
     public static RenderLayer getEntityTranslucentPartiallyTextured(Identifier textureId, float cutoutY, boolean affectsOutline) {
+        if (FabricLoader.getInstance().isModLoaded("iris")) {
+            return IrisCustomRenderLayer.getEntityTranslucentPartiallyTextured(textureId, cutoutY, affectsOutline);
+        }
         CustomGameRenderer.initRenderTypeEntityTranslucentPartiallyTexturedShader(cutoutY, MatrixStackStorage.getModelMatrixStack().peek().getPositionMatrix());
         return ENTITY_TRANSLUCENT_PARTIALLY_TEXTURED.apply(textureId, affectsOutline);
     }
 
     static {
         VOXELS = CustomGameRenderer.getRenderTypeVoxelShader().getRenderLayer(of("voxels", CustomVertexFormats.POSITION_COLOR_OVERLAY_LIGHT_NORMAL, VertexFormat.DrawMode.QUADS, 256, false, false, RenderLayer.MultiPhaseParameters.builder().program(RenderPhase.SOLID_PROGRAM).transparency(NO_TRANSPARENCY).cull(DISABLE_CULLING).lightmap(ENABLE_LIGHTMAP).overlay(ENABLE_OVERLAY_COLOR).build(true)));
-        ENTITY_TRANSLUCENT_PARTIALLY_TEXTURED = Util.memoize((id, outline) -> CustomGameRenderer.getRenderTypeEntityTranslucentPartiallyTexturedShader().getRenderLayer(RenderLayer.getEntityTranslucent(id, outline)));
+        ENTITY_TRANSLUCENT_PARTIALLY_TEXTURED = Util.memoize((id, outline) -> {
+            if (FabricLoader.getInstance().isModLoaded("iris")) {
+                return IrisCustomRenderLayer.initVoxelRenderLayer();
+            }
+            else return CustomGameRenderer.getRenderTypeEntityTranslucentPartiallyTexturedShader().getRenderLayer(RenderLayer.getEntityTranslucent(id, outline));
+        });
     }
 }
